Move httpQueries to utils
