+++ x/utils/httpQueries.go
@@ -1,4 +1,4 @@
-package coingecko
+package utils
 
 
 import(
@@ -8,7 +8,7 @@ import(
 	"io/ioutil"
 )
 
-func GetCoinGeckoReqBody(endpoint string,ptr []interface{})error{
+func QueryCoinGeckoReqBody(endpoint string,ptr []interface{})error{
 	url:="https://api.coingecko.com/api/v3/"
 	resp, err := http.Get(fmt.Sprintf("%s/%s", url, endpoint))
 	if err != nil {
